Correct cropped image preview
The HStack padding bottom is alright, there is extra padding (4) on the upload profile pic fields. It does not alter the ui though. Removed the extra padding of the bio field and added that same padding on the entire HStack so that it spaces itself correctly in correlation to the title and continue button.